refactor: organizing the files and folders
