+++ articles/IsolatedFunctions/libs/AzureServiceBusLib/Services/ServiceBusFactory.cs
@@ -1,6 +1,6 @@
 ﻿using AzureServiceBusLib.Core;
 
-namespace AzureServiceBusLib.DI;
+namespace AzureServiceBusLib.Services;
 
 internal sealed class ServiceBusFactory : IServiceBusFactory
 {
