+++ articles/IsolatedFunctions/tools/OrderPublisher.Console/Models/ISessionMessage.cs
@@ -1,4 +1,5 @@
 ﻿using AzureServiceBusLib.Core;
+using AzureServiceBusLib.Models;
 
 namespace OrderPublisher.Console.Models;
