+++ articles/IsolatedFunctions/tools/OrderPublisher.Console/Services/IOrderGenerator.cs
@@ -1,4 +1,5 @@
 ﻿using AzureServiceBusLib.Core;
+using AzureServiceBusLib.Models;
 
 namespace OrderPublisher.Console.Services;
